requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
